pyproject.toml
[tool.poetry]
name = "mlccapp"
version = "0.1.0"
description = "Multilingual Climate Change Chatbot Application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
aioboto3 = "^14.0.0"
boto3 = "^1.34.0"
cohere = "^5.11.3"
python-dotenv = "^1.0.0"
streamlit = "^1.40.1"
# torch, transformers, flagembedding REMOVED — BGE-M3 now via HF Inference API
huggingface-hub = ">=0.25.0"
langchain = "^0.3.7"
langchain-community = "^0.3.5"
langchain_cohere="^0.4.2"
pinecone-client = "^5.0.1"
redis = "^5.2.1"
tavily-python = "^0.2.1"
aiohttp = "^3.10.10"
numpy = "^1.26.4"
pandas = "^2.2.3"
tqdm = "^4.66.5"
typing-extensions = "^4.12.2"
jsonschema = "^4.23.0"
# tokenizers, opencensus, opencensus-ext-azure, opencensus-ext-logging,
# azure-storage-blob REMOVED — not imported anywhere in source code
streamlit-user-device = "^0.1.2"
streamlit-screen-stats = "^0.0.82"
streamlit-javascript = "^0.1.5"
user-agents = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.23.7"
black = "^24.2"
isort = "^5.13"
flake8 = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mlccapp = "src.main_nova:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v"
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]